requirements.txt
requests
imageio-ffmpeg
python-dotenv
